+++ src/store/getShareRoomList.ts
@@ -1,5 +1,5 @@
 import { atom } from "recoil";
-import { IShareRoomList } from "../type/shareRoom";
+import { IShareRoomList } from "../type/planShareRoom";
 
 const initial: IShareRoomList = {
   totalPages: -1,
